update build documentation command

+++ setup.py
@@ -4,7 +4,7 @@
 #   can't use setuptools yet, becouse i found data file in absolute path
 #    from setuptools import setup
 # except:
-from distutils.core import setup
+from distutils.core import setup, Command
 
 from distutils.command.build_scripts import build_scripts
 from distutils.command.build import build
@@ -35,6 +35,61 @@ def find_data_files(directory, targetFolder=""):
     return rv
 
 
+class build_doc(Command):
+    description = "build html documentation"
+    user_options = [
+        ('build-base=', 'b',
+         "base build directory (default: 'build.build-base')"),
+        ('build-doc=', 'd', "directory to build html documentation to"),
+        ('public', 'p', "build as part of public poorhttp web")
+    ]
+
+    def initialize_options(self):
+        self.build_base = None
+        self.build_doc = None
+        self.public = False
+
+    def finalize_options(self):
+        self.set_undefined_options('build',
+                                   ('build_base', 'build_base'))
+        if self.build_doc is None:
+            self.build_doc = path.join(self.build_base, 'doc')
+
+    def run(self):
+        from jinja24doc.context import Context
+        import sys
+        sys.path.insert(0, 'jinja24doc')
+        log.info("creating documentation")
+
+        def _jinja24doc(tname):
+            ctx = Context(('templates/jinja24doc', 'doc'))
+            log.info('jinja24doc processing %s ...' % tname)
+            data = ctx.generate(tname, public=self.public)
+            if not isinstance(data, str):
+                data = data.encode('utf-8')
+            with open(self.build_doc+'/'+tname, 'w') as out:
+                out.write(data)
+
+        log.info("building html documentation")
+        if self.public:
+            log.info("building as public part of poorhttp web")
+        if self.dry_run:
+            return
+        if self.build_doc is None:
+            return
+
+        self.mkpath(self.build_doc)
+        _jinja24doc('index.html')
+        _jinja24doc('tools.html')
+        _jinja24doc('templates.html')
+        _jinja24doc('api.html')
+        _jinja24doc('licence.html')
+        copyfile('templates/green.css', self.build_doc+'/style.css')
+        copyfile('templates/jinja24doc/web.css', self.build_doc+'/web.css')
+        copyfile('templates/jinja24doc//small-logo.png',
+                 self.build_doc+'/small-logo.png')
+
+
 class X_build_scripts(build_scripts):
     def run(self):
         self.mkpath('build/_scripts_')
@@ -45,45 +100,33 @@ class X_build_scripts(build_scripts):
 
 
 class X_build(build):
-    def run(self):
-        from jinja24doc.context import Context
-        import sys
-        sys.path.insert(0, 'jinja24doc')
-        log.info("creating documentation")
-
-        def _jinja24doc(tname, oname):
-            ctx = Context(('templates/jinja24doc', 'doc'))
-            log.info('jinja24doc processing %s ...' % oname)
-            data = ctx.generate(tname)
-            if not isinstance(data, str):
-                data = data.encode('utf-8')
-            with open(oname, 'w') as f:
-                f.write(data.replace('http://poorhttp.zeropage.cz/', ''))
-
-        self.mkpath('build/_html_')
-        _jinja24doc('index.html',
-                    'build/_html_/index.html')
-        _jinja24doc('jinja24doc_tools.html',
-                    'build/_html_/jinja24doc_tools.html')
-        _jinja24doc('jinja24doc_templates.html',
-                    'build/_html_/jinja24doc_templates.html')
-        _jinja24doc('jinja24doc_api.html',
-                    'build/_html_/jinja24doc_api.html')
-        _jinja24doc('jinja24doc_licence.html',
-                    'build/_html_/jinja24doc_licence.html')
-        copyfile('templates/green.css', 'build/_html_/style.css')
-        build.run(self)             # run original build
+    sub_commands = build.sub_commands + [('build_doc', lambda self: True)]
 
 
 class X_clean(clean):
+    user_options = clean.user_options + [
+        ('build-doc=', 'd', "directory to build html documentation to"),
+    ]
+
+    def initialize_options(self):
+        clean.initialize_options(self)
+        self.build_doc = None
+
+    def finalize_options(self):
+        clean.finalize_options(self)
+        if self.build_doc is None:
+            self.build_doc = path.join(self.build_base, 'doc')
+
     def run(self):
-        for directory in ('build/_html_', 'build/_scripts_'):
-            if path.exists(directory):
-                remove_tree(directory, dry_run=self.dry_run)
-            else:
-                log.warn("'%s' does not exist -- can't clean it",
-                         directory)
+        if self.all:
+            for directory in (self.build_doc, 'build/_scripts_'):
+                if path.exists(directory):
+                    remove_tree(directory, dry_run=self.dry_run)
+                else:
+                    log.warn("'%s' does not exist -- can't clean it",
+                             directory)
         clean.run(self)
+
 
 kwargs = {
     'name':             "jinja24doc",
@@ -99,7 +142,7 @@ kwargs = {
     'data_files':
         [('share/doc/jinja24doc', ['LICENCE', 'README.rst'])] +
         find_data_files('doc', 'share/doc/jinja24doc') +
-        find_data_files('build/_html_', 'share/doc/jinja24doc/html') +
+        find_data_files('build/doc', 'share/doc/jinja24doc/html') +
         find_data_files('templates', 'share/jinja24doc/templates'),
     'license':           "BSD",
     'long_description':
@@ -125,11 +168,12 @@ kwargs = {
         "Topic :: Text Processing :: Markup",
         "Topic :: Utilities"
     ],
-    'install_requires': ['jinja2 >= 2.6', 'docutils-tinyhtmlwriter'],
+    'install_requires': ['jinja2 >= 2.10', 'docutils-tinyhtmlwriter'],
     'test_suite':        'tests',
     'cmdclass':           {'build_scripts': X_build_scripts,
                            'build': X_build,
-                           'clean': X_clean},
+                           'clean': X_clean,
+                           'build_doc': build_doc},
 }
 
 setup(**kwargs)
